Changes/2525-3-5/修改了auto_clear,main,优化了了删除报告

+++ common/auto_clear.py
@@ -1,11 +1,16 @@
 import os
 import shutil
 def autoClear(n):
-    dir = f"{os.path.dirname(__file__)}/testReport/"
+    dir = f"{os.path.dirname(os.path.dirname(__file__))}/testReport/"
+    # print(f"位置：{dir}")
     file_list = os.listdir(dir)
-    print(file_list)
+    # print(file_list)
     file_list.sort(key=lambda x: os.path.getmtime(dir+x))
-    print(file_list[:n])
-    for file in file_list[:n]:
-        shutil.rmtree(dir+file)
+    # print(file_list[n:],"\n")
+    # print(len(file_list))
+    if len(file_list)/2 > n:
+        print(file_list[:n])
+        for file in file_list[:n]:
+            shutil.rmtree(dir + file)
+
